Utilisation du WS pour changer le conseiller d'un client

+++ GestiBankBrijanavi/src/com/gk/gestibank/service/impl/ConseillerService.java
@@ -1,7 +1,7 @@
 package com.gk.gestibank.service.impl;
 
 import java.text.SimpleDateFormat;
-
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,10 +16,10 @@ import com.gk.gestibank.service.IConseillerService;
 
 @Service
 public class ConseillerService implements IConseillerService {
-	
+
 	@Autowired
 	private ConseillerDao conseillerDao;
-	
+
 	@Override
 	public List<Conseiller> getAll() {
 		return conseillerDao.getAll();
@@ -32,18 +32,20 @@ public class ConseillerService implements IConseillerService {
 		conseiller.setMatricule(generateMatricule(conseiller));
 		conseillerDao.createConseiller(conseiller);
 	}
-	
-	public String generateMatricule(Conseiller conseiller) {			
-		SimpleDateFormat sdate = new SimpleDateFormat("MMYY");				
-		String matricule = conseiller.getNom().toUpperCase().charAt(0) + sdate.format(conseiller.getDateDebutContrat()) + "A";
+
+	public String generateMatricule(Conseiller conseiller) {
+		SimpleDateFormat sdate = new SimpleDateFormat("MMYY");
+		String matricule = conseiller.getNom().toUpperCase().charAt(0) + sdate.format(conseiller.getDateDebutContrat())
+				+ "A";
 		return matricule;
 	}
 
 	@Override
 	public void deleteConseiller(String matricule) {
 		// TODO : On doit vérifier si la liste de client est nulle avant de supprimer
-		// Attention : nullPointerException si matricule ne correspond à aucun conseiller
-		if(conseillerDao.getClientsFromConseiller(matricule).isEmpty()){
+		// Attention : nullPointerException si matricule ne correspond à aucun
+		// conseiller
+		if (conseillerDao.getClientsFromConseiller(matricule).isEmpty()) {
 			conseillerDao.deleteConseiller(matricule);
 		}
 	}
@@ -51,7 +53,7 @@ public class ConseillerService implements IConseillerService {
 	@Override
 	public void updateConseiller(Conseiller conseiller) {
 		// Aucune véfification : le formulaire a déjà validé les champs
-		conseillerDao.updateConseiller(conseiller);		
+		conseillerDao.updateConseiller(conseiller);
 	}
 
 	@Override
@@ -62,13 +64,24 @@ public class ConseillerService implements IConseillerService {
 	@Override
 	public void addClientToConseiller(Client client, String matricule) {
 		// vérifs ?
-		conseillerDao.addClientToConseiller(client, matricule);		
+		conseillerDao.addClientToConseiller(client, matricule);
 	}
 
 	@Override
-	public void deleteClientFromConseiller(int idClient, String matricule) {
-		// vérifs ?
-		conseillerDao.deleteClientFromConseiller(idClient, matricule);		
+	public void deleteClientFromConseiller(int idClient) {
+		// trouver le matricule du conseiller
+		String matricule = "";
+		List<Conseiller> conseillers = getAll();
+		for (int i = 0; i < conseillers.size(); i++) {
+			for (int j = 0; j < (conseillers.get(i)).getClients().size(); j++) {
+				if (conseillers.get(i).getClients().get(i).getId() == idClient) {
+					matricule = conseillers.get(i).getMatricule();
+				}
+			}
+		}
+		if(matricule != "") {
+			conseillerDao.deleteClientFromConseiller(idClient, matricule);
+		}
 	}
 
 	@Override
@@ -88,11 +101,11 @@ public class ConseillerService implements IConseillerService {
 		// vérifs ?
 		return conseillerDao.getClientsFromConseiller(matricule);
 	}
-	
-	public Conseiller getConseillerWithClient(int idClient){
-		for(Conseiller c : conseillerDao.getAll()) {
-			for(Client cl : c.getClients()) {
-				if(cl.getId() == idClient) {
+
+	public Conseiller getConseillerWithClient(int idClient) {
+		for (Conseiller c : conseillerDao.getAll()) {
+			for (Client cl : c.getClients()) {
+				if (cl.getId() == idClient) {
 					return c;
 				}
 			}
